Bid Discovery Agent Fix
We were able to set up a bid finder agent and web search sub agent using Google ADK and selenium. We are looking to perform further prompt engineering to fix issues like opening the chrome instance before the user guidance etc. Next step is to integrate with the RAG agent and download the bid documents for the next agent in the workflow.